fix(Parser): enable numeric comparison in ORDER BY
