+++ scripts/parser_test/parser_d32.ttl
@@ -1,0 +1,11 @@
+@prefix dt:   <http://example.org/datatype#> .
+@prefix ns:   <http://example.org/ns#> .
+@prefix :     <http://example.org/ns#> .
+@prefix xsd:  <http://www.w3.org/2001/XMLSchema#> .
+
+:x ns:name "Carol" .
+:x ns:height "159.5"^^xsd:double .
+:y ns:name "Alice" .
+:y ns:height "163.1"^^xsd:double .
+:z ns:name "Bob" .
+:z ns:height "171.9"^^xsd:double .
