+++ scripts/parser_test/parser_q32.sql
@@ -1,0 +1,8 @@
+prefix ns:   <http://example.org/ns#>
+
+select ?name ?h where
+{
+	?x ns:name ?name .
+	?x ns:height ?h .
+}
+order by desc(?h)
